fix: tinker menu now executes selected commands and loops back
When a command is selected in tinker, it spawns bytebell <command>
as a child process with inherited stdio, then re-opens the tinker menu.
Auto-saves any dirty config changes before running the command.
Extract tinkerHelpers.ts to stay under 300-line file size rule.
